Fix the alt and the related projects
The image src was nil so it defaulted to the alt,
which gave it an error because it was not surrounded
by quotes.  So I added quotes, and not the error doesn't show.

Also the related projects was finding all of the loan_requests
for a category, instead I just took 4.